refactor: split transforms.py into transforms/ package
Split the flat transforms.py module into a sub-package with one file
per concern (curves, patch, render). Public API is unchanged.
